+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.5.50 (Pending):
+## 0.5.50:
 * Update the AbstractDataLoaderService so that it only purges subscriptions that are "owned" by
   the data loader service and ignores any subscriptions owned by other data loaders.
 * Add EntityChangeBroker.removeAllChangeListeners() to purge listeners for a specific entity.